Cleaned up a bit + fixed Javadocs
