+++ spring-cloud-azure-appconfiguration-config/src/main/java/com/microsoft/azure/spring/cloud/config/ConfigServiceOperations.java
@@ -20,7 +20,7 @@ public interface ConfigServiceOperations {
      *
      * @param storeName the store name against which to query
      * @param queryOptions the options used to query the keys
-     * @return all key-value items {@link List< KeyValueItem >} which match given condition, result is sorted by the
+     * @return all key-value items List&lt;{@link KeyValueItem}&gt; which match given condition, result is sorted by the
      * order of the labels defined in {@code queryOptions}
      */
     List<KeyValueItem> getKeys(@NonNull String storeName, @NonNull QueryOptions queryOptions);
@@ -29,7 +29,7 @@ public interface ConfigServiceOperations {
      * List chronological/historical representation of KeyValue resource(s)
      * @param storeName the store name against which to query
      * @param queryOptions the options used to query the keys
-     * @return all key-value items {@link List< KeyValueItem >} which match given condition
+     * @return all key-value items List&lt;{@link KeyValueItem}&gt; which match given condition
      */
     List<KeyValueItem> getRevisions(@NonNull String storeName, @NonNull QueryOptions queryOptions);
 }
