+++ spring-cloud-azure-appconfiguration-config/src/test/java/com/microsoft/azure/spring/cloud/config/ConfigServiceTemplateTest.java
@@ -23,8 +23,6 @@ import com.microsoft.azure.spring.cloud.config.domain.QueryOptions;
 import com.microsoft.azure.spring.cloud.config.mock.MockCloseableHttpResponse;
 import com.microsoft.azure.spring.cloud.config.resource.ConnectionString;
 import com.microsoft.azure.spring.cloud.config.resource.ConnectionStringPool;
-import com.microsoft.azure.spring.cloud.config.stores.ConfigStore;
-
 import org.apache.http.*;
 import org.apache.http.client.methods.CloseableHttpResponse;
 import org.apache.http.client.methods.HttpUriRequest;
